+++ lib/web_translate_it.rb
@@ -8,10 +8,7 @@ module WebTranslateIt
     config = Configuration.new
     locale = I18n.locale.to_s
     return if config.ignore_locales.include?(locale)
-    puts "Looking for #{locale} translations..."
-    config.files.each do |file|
-      response_code = file.fetch(locale)
-      puts "Done. Response code: #{response_code}"
-    end
+    puts "Looking for #{locale} translations…"
+    config.files.each {|file| puts "Done. Response: #{file.fetch(locale)}" }
   end
 end