+++ spec/web_translate_it/util_spec.rb
@@ -1,0 +1,10 @@
+require File.join(File.dirname(__FILE__), '..', 'spec_helper')
+
+describe WebTranslateIt::Util do
+    
+  describe "Util.version" do
+    it "should return a String" do
+      WebTranslateIt::Util.version.should be_a(String)
+    end
+  end
+end