+++ history.md
@@ -1,4 +1,4 @@
-## Edge
+## Version 1.4.1 / 2010-01-07
 
 * Rename `rake trans:send[fr_Fr]` to `rake trans:upload[fr_FR]`
 * Remove `rake trans:version` task. Instead, version number is displayed in the welcome banner.